+++ packages/common/src/visualizations/types/index.ts
@@ -9,7 +9,6 @@ export enum VizAggregationOptions {
     AVERAGE = 'avg',
     MIN = 'min',
     MAX = 'max',
-    FIRST = 'first',
 }
 
 export const vizAggregationOptions = [
@@ -18,7 +17,6 @@ export const vizAggregationOptions = [
     VizAggregationOptions.AVERAGE,
     VizAggregationOptions.MIN,
     VizAggregationOptions.MAX,
-    VizAggregationOptions.FIRST,
 ];
 
 export const VIZ_DEFAULT_AGGREGATION = VizAggregationOptions.COUNT;