+++ packages/frontend/src/components/DataViz/visualizations/Table.tsx
@@ -2,14 +2,13 @@ import {
     type RawResultRow,
     type VizTableColumnsConfig,
 } from '@lightdash/common';
-import { Badge, Flex, Group } from '@mantine/core';
+import { Badge, Flex, Group, type FlexProps } from '@mantine/core';
 import { flexRender } from '@tanstack/react-table';
 import { SMALL_TEXT_LENGTH } from '../../common/LightTable';
 import BodyCell from '../../common/Table/ScrollableTable/BodyCell';
 import { VirtualizedArea } from '../../common/Table/ScrollableTable/TableBody';
 import {
     Table as TableStyled,
-    TableContainer,
     TABLE_HEADER_BG,
     Tr,
 } from '../../common/Table/Table.styles';
@@ -19,11 +18,13 @@ import { useTableDataModel } from '../transformers/useTableDataModel';
 type TableProps<T extends ResultsRunner> = {
     config?: VizTableColumnsConfig;
     resultsRunner: T;
+    flexProps?: FlexProps;
 };
 
 export const Table = <T extends ResultsRunner>({
     resultsRunner,
     config,
+    flexProps,
 }: TableProps<T>) => {
     const {
         tableWrapperRef,
@@ -37,110 +38,108 @@ export const Table = <T extends ResultsRunner>({
     const { headerGroups, virtualRows, rowModelRows } = getTableData();
 
     return (
-        <TableContainer $shouldExpand>
-            <Flex
-                ref={tableWrapperRef}
-                dir="column"
-                miw="100%"
-                sx={{
-                    flex: 1,
-                    overflow: 'auto',
-                }}
-            >
-                <TableStyled>
-                    <thead>
-                        <tr>
-                            {headerGroups.map((headerGroup) =>
-                                headerGroup.headers.map((header) => (
-                                    <th
-                                        key={header.id}
-                                        style={{
-                                            backgroundColor: TABLE_HEADER_BG,
-                                        }}
-                                    >
-                                        <Group spacing="two">
-                                            {config?.columns[header.id]
-                                                ?.aggregation && (
-                                                <Badge
-                                                    size="sm"
-                                                    color="indigo"
-                                                    radius="xs"
-                                                >
-                                                    {
-                                                        config?.columns[
-                                                            header.id
-                                                        ]?.aggregation
-                                                    }
-                                                </Badge>
-                                            )}
-                                            {/* TODO: do we need to check if it's a
+        <Flex
+            ref={tableWrapperRef}
+            direction="column"
+            miw="100%"
+            {...flexProps}
+            sx={{
+                overflow: 'auto',
+                fontFamily: "'Inter', sans-serif",
+                fontFeatureSettings: "'tnum'",
+                flexGrow: 1,
+                ...flexProps?.sx,
+            }}
+        >
+            <TableStyled>
+                <thead>
+                    <tr>
+                        {headerGroups.map((headerGroup) =>
+                            headerGroup.headers.map((header) => (
+                                <th
+                                    key={header.id}
+                                    style={{
+                                        backgroundColor: TABLE_HEADER_BG,
+                                    }}
+                                >
+                                    <Group spacing="two">
+                                        {config?.columns[header.id]
+                                            ?.aggregation && (
+                                            <Badge
+                                                size="sm"
+                                                color="indigo"
+                                                radius="xs"
+                                            >
+                                                {
+                                                    config?.columns[header.id]
+                                                        ?.aggregation
+                                                }
+                                            </Badge>
+                                        )}
+                                        {/* TODO: do we need to check if it's a
                                             placeholder? */}
-                                            {flexRender(
-                                                header.column.columnDef.header,
-                                                header.getContext(),
-                                            )}
-                                        </Group>
-                                    </th>
-                                )),
-                            )}
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {paddingTop > 0 && (
-                            <VirtualizedArea
-                                cellCount={columnsCount}
-                                padding={paddingTop}
-                            />
+                                        {flexRender(
+                                            header.column.columnDef.header,
+                                            header.getContext(),
+                                        )}
+                                    </Group>
+                                </th>
+                            )),
                         )}
-                        {virtualRows.map(({ index }) => {
-                            return (
-                                <Tr key={index} $index={index}>
-                                    {rowModelRows[index]
-                                        .getVisibleCells()
-                                        .map((cell) => {
-                                            const cellValue =
-                                                cell.getValue() as
-                                                    | RawResultRow[0]
-                                                    | undefined;
+                    </tr>
+                </thead>
+                <tbody>
+                    {paddingTop > 0 && (
+                        <VirtualizedArea
+                            cellCount={columnsCount}
+                            padding={paddingTop}
+                        />
+                    )}
+                    {virtualRows.map(({ index }) => {
+                        return (
+                            <Tr key={index} $index={index}>
+                                {rowModelRows[index]
+                                    .getVisibleCells()
+                                    .map((cell) => {
+                                        const cellValue = cell.getValue() as
+                                            | RawResultRow[0]
+                                            | undefined;
 
-                                            return (
-                                                <BodyCell
-                                                    key={cell.id}
-                                                    index={index}
-                                                    cell={cell}
-                                                    isNumericItem={false}
-                                                    hasData={!!cellValue}
-                                                    isLargeText={
-                                                        (
-                                                            cellValue?.toString() ||
-                                                            ''
-                                                        ).length >
-                                                        SMALL_TEXT_LENGTH
-                                                    }
-                                                >
-                                                    {cell.getIsPlaceholder()
-                                                        ? null
-                                                        : flexRender(
-                                                              cell.column
-                                                                  .columnDef
-                                                                  .cell,
-                                                              cell.getContext(),
-                                                          )}
-                                                </BodyCell>
-                                            );
-                                        })}
-                                </Tr>
-                            );
-                        })}
-                        {paddingBottom > 0 && (
-                            <VirtualizedArea
-                                cellCount={columnsCount}
-                                padding={paddingBottom}
-                            />
-                        )}
-                    </tbody>
-                </TableStyled>
-            </Flex>
-        </TableContainer>
+                                        return (
+                                            <BodyCell
+                                                key={cell.id}
+                                                index={index}
+                                                cell={cell}
+                                                isNumericItem={false}
+                                                hasData={!!cellValue}
+                                                isLargeText={
+                                                    (
+                                                        cellValue?.toString() ||
+                                                        ''
+                                                    ).length > SMALL_TEXT_LENGTH
+                                                }
+                                            >
+                                                {cell.getIsPlaceholder()
+                                                    ? null
+                                                    : flexRender(
+                                                          cell.column.columnDef
+                                                              .cell,
+                                                          cell.getContext(),
+                                                      )}
+                                            </BodyCell>
+                                        );
+                                    })}
+                            </Tr>
+                        );
+                    })}
+                    {paddingBottom > 0 && (
+                        <VirtualizedArea
+                            cellCount={columnsCount}
+                            padding={paddingBottom}
+                        />
+                    )}
+                </tbody>
+            </TableStyled>
+        </Flex>
     );
 };