+++ packages/frontend/src/hooks/dashboard/useDashboard.ts
@@ -245,6 +245,7 @@ export const useUpdateDashboard = (
                     'dashboards-containing-chart',
                 ]);
                 await queryClient.resetQueries(['saved_dashboard_query', id]);
+                await queryClient.invalidateQueries(['content']);
                 const onlyUpdatedName: boolean =
                     Object.keys(variables).length === 1 &&
                     Object.keys(variables).includes('name');
@@ -294,6 +295,7 @@ export const useMoveDashboardMutation = () => {
                 await queryClient.invalidateQueries([
                     'most-popular-and-recently-updated',
                 ]);
+                await queryClient.invalidateQueries(['content']);
                 queryClient.setQueryData(
                     ['saved_dashboard_query', data.uuid],
                     data,
@@ -395,6 +397,7 @@ export const useDuplicateDashboardMutation = (
                 await queryClient.invalidateQueries([
                     'most-popular-and-recently-updated',
                 ]);
+                await queryClient.invalidateQueries(['content']);
                 showToastSuccess({
                     title: `Dashboard successfully duplicated!`,
                     action: options?.showRedirectButton
@@ -434,6 +437,7 @@ export const useDashboardDeleteMutation = () => {
             await queryClient.invalidateQueries([
                 'most-popular-and-recently-updated',
             ]);
+            await queryClient.invalidateQueries(['content']);
             showToastSuccess({
                 title: `Deleted! Dashboard was deleted.`,
             });