+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+# [0.1239.0](https://github.com/lightdash/lightdash/compare/0.1238.3...0.1239.0) (2024-09-03)
+
+
+### Features
+
+* semantic viewer layout ([#11356](https://github.com/lightdash/lightdash/issues/11356)) ([bf16d81](https://github.com/lightdash/lightdash/commit/bf16d814ba1351f3187198cf2d1b3e7606233d28))
+
+## [0.1238.3](https://github.com/lightdash/lightdash/compare/0.1238.2...0.1238.3) (2024-09-03)
+
+
+### Bug Fixes
+
+* self-hosted script to override password ([#11317](https://github.com/lightdash/lightdash/issues/11317)) ([0b3d033](https://github.com/lightdash/lightdash/commit/0b3d033cd79cb8b7266334e6d70bc2d5db48e030))
+
+## [0.1238.2](https://github.com/lightdash/lightdash/compare/0.1238.1...0.1238.2) (2024-09-03)
+
+
+### Bug Fixes
+
+* display sql runner error ([#11357](https://github.com/lightdash/lightdash/issues/11357)) ([3534f3a](https://github.com/lightdash/lightdash/commit/3534f3aeaf9afab1fdcef479645a1ebc0bb56959))
+
+## [0.1238.1](https://github.com/lightdash/lightdash/compare/0.1238.0...0.1238.1) (2024-09-03)
+
+
+### Bug Fixes
+
+* fixed delete dashboards from resources list ([#11205](https://github.com/lightdash/lightdash/issues/11205)) ([f3a4d48](https://github.com/lightdash/lightdash/commit/f3a4d4848e30c17b58b71c73c87c6f4337c6da8f))
+* only push if not included ([#11352](https://github.com/lightdash/lightdash/issues/11352)) ([076ad46](https://github.com/lightdash/lightdash/commit/076ad46e5edf929be571ee184babce41e93c3f82))
+
+# [0.1238.0](https://github.com/lightdash/lightdash/compare/0.1237.4...0.1238.0) (2024-09-02)
+
+
+### Features
+
+* allow fine tuning query limits for non-table visualizations ([#11321](https://github.com/lightdash/lightdash/issues/11321)) ([ac9f01f](https://github.com/lightdash/lightdash/commit/ac9f01fb5eca4c7927ba5a5b28f6a909dbe122f8))
+
+## [0.1237.4](https://github.com/lightdash/lightdash/compare/0.1237.3...0.1237.4) (2024-09-02)
+
+
+### Bug Fixes
+
+* dashboard mentions not working sometimes and added loading state ([#11246](https://github.com/lightdash/lightdash/issues/11246)) ([1dd45b0](https://github.com/lightdash/lightdash/commit/1dd45b035b9485c6f646f68a4a4a5520eca15010))
+
+## [0.1237.3](https://github.com/lightdash/lightdash/compare/0.1237.2...0.1237.3) (2024-09-02)
+
+
+### Bug Fixes
+
+* handle case sensitivity for SQL catalog for snowflake ([#11315](https://github.com/lightdash/lightdash/issues/11315)) ([ad268a7](https://github.com/lightdash/lightdash/commit/ad268a722590731c1b1c28551bc214b32bda4ffc)), closes [#11134](https://github.com/lightdash/lightdash/issues/11134) [#11134](https://github.com/lightdash/lightdash/issues/11134) [#11134](https://github.com/lightdash/lightdash/issues/11134)
+* use <= for row_index limit ([#11343](https://github.com/lightdash/lightdash/issues/11343)) ([98444ba](https://github.com/lightdash/lightdash/commit/98444ba4eeecad1e5006ab1cf25c869a67981828))
+
+## [0.1237.2](https://github.com/lightdash/lightdash/compare/0.1237.1...0.1237.2) (2024-09-02)
+
+
+### Bug Fixes
+
+* remove semantic viewer aggregation functions ([#11333](https://github.com/lightdash/lightdash/issues/11333)) ([e4e8520](https://github.com/lightdash/lightdash/commit/e4e8520615ecdfe259342725fd61b8074f5b3761))
+
 ## [0.1237.1](https://github.com/lightdash/lightdash/compare/0.1237.0...0.1237.1) (2024-09-02)